[backprop] Implement a first neural network for classifying handwritten digits.
Implementation is based on the theory of nets written in the Michael Nielsen book (http://neuralnetworksanddeeplearning.com).
Before using the data has to be extracted to "1_backprop\backprop_cpp\data" folder.

The implementation was checked against python one and it gives approx. the same results as the reference (python) implementation.